moving-validator-config-a-separate-file
