+++ src/ui-forms/textField.jsx
@@ -1,10 +1,14 @@
 import React from "react";
+import styles from "./ui-form.module.css";
 
 const TextField = ({ label, type, name, onChange, value, error }) => {
   return (
     <div>
-      <label htmlFor={name}>{label}</label>
+      <label htmlFor={name} className={styles.label_name}>
+        {label}
+      </label>
       <input
+        className={styles.input}
         type={type}
         id={name}
         value={value.name}
