+++ src/ui-forms/ui-form.module.css
@@ -1,5 +1,6 @@
 @import url("https://fonts.googleapis.com/css2?family=Roboto&display=swap");
 .main_div {
+  font-family: "Roboto", sans-serif;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -25,10 +26,13 @@
 }
 
 .button_submit {
+  display: flex;
+  align-items: center;
+  justify-content: center;
   border: 1px solid black;
   border-radius: 15px;
   width: auto;
-  margin: 15px 0 0 87px;
+  margin: 15px 0 0 105px;
   width: 100px;
 }
